* Bitmanipulation von MCP nach BaseIO verschoben * OLED Library SPI Geschwindigkeit auf max. 32MHz erhöht * OLED Zugriffsoptimierungen + Timer Kommentare * Timer Fehlerbehebungen

+++ src/Timer.cpp
@@ -8,7 +8,7 @@ Timer::Timer(unsigned long intervall)
 void Timer::Update()
 {
     // Prüfen ob der Timer läuft und abgelaufen ist
-    if(IsRunning && millis() - StartTime >= (unsigned long)Intervall)
+    if(IsRunning && millis() - StartTime >= Intervall)
     {
         // Timer ist abgelaufen
         Stop();
